+++ src/components/domain/Page/index.ts
@@ -1,1 +1,2 @@
 export { DeletePageModal } from './DeletePageModal';
+export { PageCountupCard } from './PageCountupCard';
